src/requirements.txt
aiogram~=3.4
python-dotenv~=1.0
aiosqlite~=0.20
